move SofaTracerRCFBeanPostProcessor to redis module
